+++ massiv-io/src/Data/Massiv/Array/IO/Base.hs
@@ -1,8 +1,10 @@
 {-# LANGUAGE FlexibleContexts      #-}
+{-# LANGUAGE FlexibleInstances     #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE RankNTypes            #-}
 {-# LANGUAGE ScopedTypeVariables   #-}
 {-# LANGUAGE TypeFamilies          #-}
+{-# LANGUAGE UndecidableInstances  #-}
 -- |
 -- Module      : Data.Massiv.Array.IO.Base
 -- Copyright   : (c) Alexey Kuleshevich 2018
@@ -29,11 +31,11 @@ module Data.Massiv.Array.IO.Base
   , convertEither
   ) where
 
-import           Control.Exception    (throw, Exception)
-import           Data.Massiv.Array    (Array, Ix2)
+import           Control.Exception    (Exception, throw)
 import qualified Data.ByteString      as B (ByteString)
 import qualified Data.ByteString.Lazy as BL (ByteString)
 import           Data.Default         (Default (..))
+import           Data.Massiv.Array
 import           Data.Maybe           (fromMaybe)
 import           Data.Typeable
 import           Graphics.ColorSpace  (ColorSpace, Pixel)
@@ -182,3 +184,4 @@ convertEither f showCS conv eImg =
         showsTypeRep (typeRep (Proxy :: Proxy e)) ">"))
     Right
     (conv eImg)
+