+++ massiv-io/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.1.3
+
+* Instance of `Writable` for `DW` arrays.
+
 # 0.1.2
 
 * Exposed `Elevator` internal functions.